requirements.txt
python-telegram-bot
aiohttp
aiosqlite
python-dotenv
praw
tcp_latency
google-generativeai
PyYAML